Rename to functionaltests
